+++ libraries/session/src/main/java/androidx/media3/session/MediaSessionImpl.java
@@ -66,6 +66,7 @@ import androidx.annotation.Nullable;
 import androidx.concurrent.futures.ResolvableFuture;
 import androidx.media3.common.AudioAttributes;
 import androidx.media3.common.DeviceInfo;
+import androidx.media3.common.Flags;
 import androidx.media3.common.MediaItem;
 import androidx.media3.common.MediaLibraryInfo;
 import androidx.media3.common.MediaMetadata;
@@ -195,7 +196,8 @@ import org.checkerframework.checker.initialization.qual.Initialized;
             + MediaLibraryInfo.VERSION_SLASHY
             + "] ["
             + Util.DEVICE_DEBUG_INFO
-            + "]");
+            + "] "
+            + Flags.getLogString());
     this.instance = instance;
     this.context = context;
     sessionId = id;